mds restarts: only restart if necessary
This commit adds module and runner functions to distinguish of the
binary or config changed and chooses the appropriate restart method for
mds daemons. As a side effect this short circuits the restart checks in
case non is needed which should speed up things if applied to other
daemons as well.